+++ spinta/datasets/commands/link.py
@@ -50,10 +50,6 @@ def link(context: Context, entity: Entity):
     else:
         entity.dataset = None
 
-    # XXX: https://gitlab.com/atviriduomenys/spinta/-/issues/44
-    pkeys: List[str] = entity.pkeys or []
-    entity.pkeys = [entity.model.properties[k] for k in pkeys]
-
 
 @commands.link.register(Context, Attribute)
 def link(context: Context, attribute: Attribute):