+++ spinta/types/ref/link.py
@@ -32,7 +32,7 @@ def link(context: Context, dtype: Ref) -> None:
                 raise ModelReferenceKeyNotFound(dtype, ref=rprop, model=dtype.model)
             refprops.append(dtype.model.properties[rprop])
         dtype.refprops = refprops
+    elif dtype.model.external:
+        dtype.refprops = [*dtype.model.external.pkeys]
     else:
-        dtype.refprops = [
-            dtype.model.properties['_id'],
-        ]
+        dtype.refprops = [dtype.model.properties['_id']]